pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "quickrefs-index"
version = "1.0"
authors = [
  { name="Parantapa Bhattacharya", email="pb+pypi@parantapa.net" },
]
description = "Quick reference index."
readme = "README.rst"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    "click",
    "python-dateutil",
    "pydantic",
]

[project.urls]
"Homepage" = "https://github.com/parantapa/quickrefs-index"

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
quickrefs-index = "quickrefs_index:quickrefs_index"
